use long for evaluation close #17

+++ src/main/java/cryptator/solver/gentest/CryptaGTSolver.java
@@ -174,7 +174,7 @@ public class CryptaGTSolver {
             }
         }
         ICryptaEvaluation chk = new CryptaEvaluation();
-        int v;
+        long v;
         try {
             v = chk.evaluate(root, new CryptaGTSolution(map), 10);
         } catch (CryptaEvaluationException e) {
